requirements.txt
labelbox==3.42.0
Pillow==9.4.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
tqdm==4.65.0
